touch script gentoo2ebuilds to signal creation
With previous automated commit 280734a4 a script has been added that
help creating this repository from the upstream gentoo one

gentoo2ebuilds is a very rough script that replace the previous method
to create our little ebuild repo.
Basically it has a list of categories and packages we need and copy
those from upstream.

It replace a never working script that tried to keep git history